Cleanup: move to new encoding file
