+++ services/rfq/relayer/pricer/fee_pricer.go
@@ -3,7 +3,6 @@ package pricer
 
 import (
 	"context"
-	"encoding/binary"
 	"fmt"
 	"math/big"
 	"strings"
@@ -16,6 +15,7 @@ import (
 	"github.com/synapsecns/sanguine/core/metrics"
 	"github.com/synapsecns/sanguine/ethergo/submitter"
 	"github.com/synapsecns/sanguine/services/rfq/contracts/fastbridgev2"
+	"github.com/synapsecns/sanguine/services/rfq/relayer/chain"
 	"github.com/synapsecns/sanguine/services/rfq/relayer/relconfig"
 	"github.com/synapsecns/sanguine/services/rfq/relayer/reldb"
 	"go.opentelemetry.io/otel/attribute"
@@ -233,7 +233,7 @@ func (f *feePricer) getZapGasEstimate(ctx context.Context, destination uint32, q
 		fastBridgeV2ABI = &parsedABI
 	}
 
-	rawRequest, err := EncodeQuoteRequest(quoteRequest.Transaction)
+	rawRequest, err := chain.EncodeQuoteRequest(quoteRequest.Transaction)
 	if err != nil {
 		return 0, fmt.Errorf("could not encode quote data: %w", err)
 	}
@@ -260,81 +260,6 @@ func (f *feePricer) getZapGasEstimate(ctx context.Context, destination uint32, q
 	}
 
 	return gasEstimate, nil
-}
-
-const (
-	// Field sizes in bytes
-	sizeVersion = 2
-	sizeChainID = 4
-	sizeAddress = 20
-	sizeUint256 = 32
-
-	// Field offsets in bytes
-	offsetVersion            = 0
-	offsetOriginChainID      = offsetVersion + sizeVersion
-	offsetDestChainID        = offsetOriginChainID + sizeChainID
-	offsetOriginSender       = offsetDestChainID + sizeChainID
-	offsetDestRecipient      = offsetOriginSender + sizeAddress
-	offsetOriginToken        = offsetDestRecipient + sizeAddress
-	offsetDestToken          = offsetOriginToken + sizeAddress
-	offsetOriginAmount       = offsetDestToken + sizeAddress
-	offsetDestAmount         = offsetOriginAmount + sizeUint256
-	offsetOriginFeeAmount    = offsetDestAmount + sizeUint256
-	offsetDeadline           = offsetOriginFeeAmount + sizeUint256
-	offsetNonce              = offsetDeadline + sizeUint256
-	offsetExclusivityRelayer = offsetNonce + sizeUint256
-	offsetExclusivityEndTime = offsetExclusivityRelayer + sizeAddress
-	offsetZapNative          = offsetExclusivityEndTime + sizeUint256
-	offsetZapData            = offsetZapNative + sizeUint256
-)
-
-// Helper function to properly encode uint256
-func padUint256(b *big.Int) []byte {
-	// Convert big.Int to bytes
-	bytes := b.Bytes()
-	// Create 32-byte array (initialized to zeros)
-	result := make([]byte, 32)
-	// Copy bytes to right side of array (left-pad with zeros)
-	copy(result[32-len(bytes):], bytes)
-	return result
-}
-
-// EncodeQuoteRequest encodes a quote request into a byte array.
-func EncodeQuoteRequest(tx fastbridgev2.IFastBridgeV2BridgeTransactionV2) ([]byte, error) {
-	result := make([]byte, offsetZapData)
-
-	// Version
-	result[offsetVersion] = 0
-	result[offsetVersion+1] = 2
-
-	// Chain IDs
-	binary.BigEndian.PutUint32(result[offsetOriginChainID:offsetOriginChainID+sizeChainID], tx.OriginChainId)
-	binary.BigEndian.PutUint32(result[offsetDestChainID:offsetDestChainID+sizeChainID], tx.DestChainId)
-
-	// Addresses
-	copy(result[offsetOriginSender:offsetOriginSender+sizeAddress], tx.OriginSender.Bytes())
-	copy(result[offsetDestRecipient:offsetDestRecipient+sizeAddress], tx.DestRecipient.Bytes())
-	copy(result[offsetOriginToken:offsetOriginToken+sizeAddress], tx.OriginToken.Bytes())
-	copy(result[offsetDestToken:offsetDestToken+sizeAddress], tx.DestToken.Bytes())
-
-	// uint256 values
-	copy(result[offsetOriginAmount:offsetOriginAmount+sizeUint256], padUint256(tx.OriginAmount))
-	copy(result[offsetDestAmount:offsetDestAmount+sizeUint256], padUint256(tx.DestAmount))
-	copy(result[offsetOriginFeeAmount:offsetOriginFeeAmount+sizeUint256], padUint256(tx.OriginFeeAmount))
-	copy(result[offsetDeadline:offsetDeadline+sizeUint256], padUint256(tx.Deadline))
-	copy(result[offsetNonce:offsetNonce+sizeUint256], padUint256(tx.Nonce))
-
-	// Exclusivity address
-	copy(result[offsetExclusivityRelayer:offsetExclusivityRelayer+sizeAddress], tx.ExclusivityRelayer.Bytes())
-
-	// More uint256 values
-	copy(result[offsetExclusivityEndTime:offsetExclusivityEndTime+sizeUint256], padUint256(tx.ExclusivityEndTime))
-	copy(result[offsetZapNative:offsetZapNative+sizeUint256], padUint256(tx.ZapNative))
-
-	// Append ZapData
-	result = append(result, tx.ZapData...)
-
-	return result, nil
 }
 
 func (f *feePricer) GetTotalFee(parentCtx context.Context, origin, destination uint32, denomToken string, isQuote bool, quoteRequest *reldb.QuoteRequest) (_ *big.Int, err error) {
